pyproject.toml
[tool.poetry]
name = "hub"
version = "0.1.0"
description = ""
authors = ["Alchemmist <anton.ingrish@gmail.com>"]
license = "gpl"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
flask = "^2.3.3"
pyyaml = "^6.0.1"
flask-login = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
